+++ Changes.md
@@ -13,6 +13,7 @@
 
 * ehlo_hello_message: config/ehlo_hello_message can be used to overwrite the EHLO/HELO msg replacing `, Haraka is at your service` #2498
 
+* connection: add connection.remote.is_local flag for detecting loopback and link local IPs
 
 ## 2.8.21 - Jul 20, 2018